+++ src/ppl/nn/engines/x86/impls/src/ppl/kernel/x86/fp32/resize2d/resize2d_fp32_avx.cpp
@@ -83,7 +83,7 @@ ppl::common::RetCode resize2d_n16chw_pytorch_2linear_floor_fp32_avx(
                 h0_lambda = 1.0f - h1_lambda;
             }
 
-            for (int ow = 0; ow < dst_w; ++ow) {
+            for (int64_t ow = 0; ow < dst_w; ++ow) {
                 int64_t w0, w1;
                 float w0_lambda, w1_lambda;
                 w0        = w0_vec[ow];
